move testToEquipmentInfos to EquipmentInfosServiceTests
