+++ src/test/java/com/powsybl/network/conversion/server/EquipmentInfosServiceTests.java
@@ -10,10 +10,7 @@ import com.google.common.collect.Iterables;
 import com.powsybl.commons.datasource.ReadOnlyDataSource;
 import com.powsybl.commons.datasource.ResourceDataSource;
 import com.powsybl.commons.datasource.ResourceSet;
-import com.powsybl.iidm.network.Identifiable;
-import com.powsybl.iidm.network.IdentifiableType;
-import com.powsybl.iidm.network.Network;
-import com.powsybl.iidm.network.VariantManagerConstants;
+import com.powsybl.iidm.network.*;
 import com.powsybl.iidm.xml.XMLImporter;
 import com.powsybl.network.conversion.server.dto.EquipmentInfos;
 import com.powsybl.network.conversion.server.dto.SubstationInfos;
@@ -50,6 +47,9 @@ public class EquipmentInfosServiceTests {
     @Autowired
     private EquipmentInfosService equipmentInfosService;
 
+    @Autowired
+    private NetworkConversionService networkConversionService;
+
     @Before
     public void setup() {
         equipmentInfosService.deleteAllOnInitialVariant(NETWORK_UUID);
@@ -71,6 +71,149 @@ public class EquipmentInfosServiceTests {
 
         equipmentInfosService.deleteAllOnInitialVariant(NETWORK_UUID);
         assertEquals(0, Iterables.size(equipmentInfosService.findAll(NETWORK_UUID)));
+    }
+
+    @Test
+    public void testToEquipmentInfos() {
+        ReadOnlyDataSource dataSource = new ResourceDataSource("testCase", new ResourceSet("", "testCase.xiidm"));
+        Network network = new XMLImporter().importData(dataSource, new NetworkFactoryImpl(), null);
+        UUID networkUuid = UUID.randomUUID();
+
+        VoltageLevel vl = network.getVoltageLevel("BBE1AA1");
+        EquipmentInfos equipmentInfos = networkConversionService.toEquipmentInfos(vl, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        EquipmentInfos expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA1")
+                .name("BBE1AA1")
+                .type(IdentifiableType.VOLTAGE_LEVEL.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("BBE1AA1").name("BBE1AA1").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        Substation substation = network.getSubstation("BBE1AA");
+        equipmentInfos = networkConversionService.toEquipmentInfos(substation, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA")
+                .name("BBE1AA")
+                .type(IdentifiableType.SUBSTATION.name())
+                .voltageLevels(Set.of(
+                        VoltageLevelInfos.builder().id("BBE1TR1").name("BBE1TR1").build(),
+                        VoltageLevelInfos.builder().id("BBE1TR2").name("BBE1TR2").build(),
+                        VoltageLevelInfos.builder().id("BBE1TR3").name("BBE1TR3").build(),
+                        VoltageLevelInfos.builder().id("BBE1AA2").name("BBE1AA2").build(),
+                        VoltageLevelInfos.builder().id("BBE1AA1").name("BBE1AA1").build(),
+                        VoltageLevelInfos.builder().id("BBE1AA5").name("BBE1AA5").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        Switch switch1 = network.getSwitch("FRA1AA1_switch");
+        equipmentInfos = networkConversionService.toEquipmentInfos(switch1, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("FRA1AA1_switch")
+                .name("FRA1AA1_switch")
+                .type(IdentifiableType.SWITCH.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("FRA1AA1").name("FRA1AA1").build()))
+                .substations(Set.of(SubstationInfos.builder().id("FRA1AA").name("FRA1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        Load load = network.getLoad("BBE1AA1 _load");
+        equipmentInfos = networkConversionService.toEquipmentInfos(load, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA1 _load")
+                .name("BBE1AA1 _load")
+                .type(IdentifiableType.LOAD.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("BBE1AA1").name("BBE1AA1").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        Bus bus = network.getBusBreakerView().getBus("BBE1AA1 ");
+        equipmentInfos = networkConversionService.toEquipmentInfos(bus, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA1 ")
+                .name("BBE1AA1 ")
+                .type(IdentifiableType.BUS.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("BBE1AA1").name("BBE1AA1").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        Generator generator = network.getGenerator("BBE1AA1 _generator");
+        equipmentInfos = networkConversionService.toEquipmentInfos(generator, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA1 _generator")
+                .name("BBE1AA1 _generator")
+                .type(IdentifiableType.GENERATOR.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("BBE1AA1").name("BBE1AA1").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        HvdcLine hvdcLine = network.getHvdcLine("FRA1AA_BBE1AA_hvdcline");
+        equipmentInfos = networkConversionService.toEquipmentInfos(hvdcLine, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("FRA1AA_BBE1AA_hvdcline")
+                .name("FRA1AA_BBE1AA_hvdcline")
+                .type(IdentifiableType.HVDC_LINE.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("FRA1AA1").name("FRA1AA1").build(), VoltageLevelInfos.builder().id("BBE1AA5").name("BBE1AA5").build()))
+                .substations(Set.of(SubstationInfos.builder().id("FRA1AA").name("FRA1AA").build(), SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        TwoWindingsTransformer twoWindingsTransformer = network.getTwoWindingsTransformer("BBE1AA2  BBE3AA1  2");
+        equipmentInfos = networkConversionService.toEquipmentInfos(twoWindingsTransformer, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA2  BBE3AA1  2")
+                .name("BBE1AA2  BBE3AA1  2")
+                .type(IdentifiableType.TWO_WINDINGS_TRANSFORMER.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("BBE1AA2").name("BBE1AA2").build(), VoltageLevelInfos.builder().id("BBE1AA1").name("BBE1AA1").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        ThreeWindingsTransformer threeWindingsTransformer = network.getThreeWindingsTransformer("BBE1AA_w3t");
+        equipmentInfos = networkConversionService.toEquipmentInfos(threeWindingsTransformer, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA_w3t")
+                .name("BBE1AA_w3t")
+                .type(IdentifiableType.THREE_WINDINGS_TRANSFORMER.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("BBE1TR1").name("BBE1TR1").build(), VoltageLevelInfos.builder().id("BBE1TR2").name("BBE1TR2").build(), VoltageLevelInfos.builder().id("BBE1TR3").name("BBE1TR3").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
+
+        Line line = network.getLine("BBE1AA1  BBE2AA1  1");
+        equipmentInfos = networkConversionService.toEquipmentInfos(line, networkUuid, VariantManagerConstants.INITIAL_VARIANT_ID);
+        expectedEquipmentInfos = EquipmentInfos.builder()
+                .networkUuid(networkUuid)
+                .variantId(VariantManagerConstants.INITIAL_VARIANT_ID)
+                .id("BBE1AA1  BBE2AA1  1")
+                .name("BBE1AA1  BBE2AA1  1")
+                .type(IdentifiableType.LINE.name())
+                .voltageLevels(Set.of(VoltageLevelInfos.builder().id("BBE1AA1").name("BBE1AA1").build(), VoltageLevelInfos.builder().id("BBE2AA1").name("BBE2AA1").build()))
+                .substations(Set.of(SubstationInfos.builder().id("BBE1AA").name("BBE1AA").build(), SubstationInfos.builder().id("BBE2AA").name("BBE2AA").build()))
+                .build();
+        assertEquals(expectedEquipmentInfos, equipmentInfos);
     }
 
     @Test
